PullDemoUseCase refactor
