+++ app/src/main/java/org/eyeseetea/malariacare/data/database/iomodules/dhis/importer/PullDemoController.java
@@ -2,21 +2,9 @@ package org.eyeseetea.malariacare.data.database.iomodules.dhis.importer;
 
 import android.content.Context;
 import android.content.res.AssetManager;
-import android.net.Uri;
 import android.util.Log;
-
-import com.raizlabs.android.dbflow.config.DHIS2GeneratedDatabaseHolder;
-import com.raizlabs.android.dbflow.config.DatabaseDefinition;
-import com.raizlabs.android.dbflow.config.EyeSeeTeaGeneratedDatabaseHolder;
-import com.raizlabs.android.dbflow.config.FlowConfig;
-import com.raizlabs.android.dbflow.config.FlowManager;
-
-import org.eyeseetea.malariacare.data.database.AppDatabase;
-import org.eyeseetea.malariacare.data.database.datasources.ConversionLocalDataSource;
-import org.eyeseetea.malariacare.data.database.utils.PopulateDB;
+import org.eyeseetea.malariacare.data.database.datasources.DBFlowLocalDataSource;
 import org.eyeseetea.malariacare.domain.boundary.IPullDemoController;
-import org.eyeseetea.sdk.common.DatabaseUtils;
-import org.eyeseetea.sdk.common.FileUtils;
 
 import java.io.IOException;
 import java.io.InputStream;
@@ -26,10 +14,12 @@ public class PullDemoController implements IPullDemoController {
 
     final String DATABASE_FOLDER = "database/";
     final String DATABASE_FILE = "EyeSeeTeaDB.db";
+    DBFlowLocalDataSource dbFlowLocalDataSource;
     Context context;
 
     public PullDemoController(Context context) {
         this.context = context;
+        dbFlowLocalDataSource = new DBFlowLocalDataSource(context);
     }
 
     @Override
@@ -45,69 +35,14 @@ public class PullDemoController implements IPullDemoController {
         }
         try {
             if (inputStream != null) {
-                pullFromDB(context, inputStream);
+                dbFlowLocalDataSource.pullFromDB(context, inputStream);
             } else {
-                pullFromCsv(context);
+                dbFlowLocalDataSource.pullFromCsv(context);
             }
             callback.onComplete();
         } catch (IOException e) {
             e.printStackTrace();
             callback.onError(e);
         }
-    }
-
-    private void pullFromCsv(Context context) throws IOException {
-        ConversionLocalDataSource.wipeDataBase();
-        deleteSQLiteMetadata();
-        Log.d(TAG, "Populate from csv start");
-        populateFromDB(context);
-        Log.d(TAG, "Populate from csv finished");
-    }
-
-    private void pullFromDB(Context context,
-            InputStream inputStream) throws IOException {
-        Log.d(TAG, "Copy Database from assets started");
-        FlowManager.destroy();
-        copyDBFromFile(inputStream);
-        reinitializeDbFlowDatabases(context);
-        Log.d(TAG, "Copy Database from assets finished");
-    }
-
-    public void copyDBFromFile(InputStream inputStream)
-            throws IOException {
-        FileUtils.copyInputStreamToFile(inputStream, DatabaseUtils.getAppDatabaseFile(AppDatabase.NAME, context.getPackageName()));
-    }
-
-    public void populateFromDB(Context context) throws IOException {
-        PopulateDB.populateDB(context.getAssets());
-    }
-
-    /**
-     * This method removes the sqlite_sequence table that contains the last autoincrement value for
-     * each table
-     */
-    private void deleteSQLiteMetadata() {
-        String sqlCopy = "Delete from sqlite_sequence";
-        DatabaseDefinition databaseDefinition =
-                FlowManager.getDatabase(AppDatabase.class);
-        databaseDefinition.getWritableDatabase().execSQL(sqlCopy);
-
-    }
-
-    /**
-     * This method reinitialize the DBFlow configurations to make DBFlow work in the appDB and in
-     * the SDK.
-     */
-    public void reinitializeDbFlowDatabases(Context context) {
-        FlowConfig flowConfig = new FlowConfig
-                .Builder(context)
-                .addDatabaseHolder(EyeSeeTeaGeneratedDatabaseHolder.class)
-                .build();
-        FlowManager.init(flowConfig);
-        FlowConfig flowConfigDhis = new FlowConfig
-                .Builder(context)
-                .addDatabaseHolder(DHIS2GeneratedDatabaseHolder.class)
-                .build();
-        FlowManager.init(flowConfigDhis);
     }
 }
